+++ frontend/lib/util/generate_id_util.dart
@@ -1,0 +1,5 @@
+class GenerateIdUtil {
+  static String generateId(String lastId) {
+    return lastId.isEmpty ? '1' : (int.parse(lastId) + 1).toString();
+  }
+}
